# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SOS App — Backend API",
    "version": "0.3.0",
    "description": "Snap or Slap challenge app. Sprint 3: Friends & Social System. All endpoints return JSON. No auth — use ?user_id=UUID."
  },
  "servers": [
    { "url": "https://backend-production-2ba1.up.railway.app", "description": "Railway Production" },
    { "url": "http://localhost:3000", "description": "Local Development" }
  ],
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user profile",
        "description": "Get profile by user_id query param. Test in browser: /api/users/me?user_id=UUID",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "User UUID"
          }
        ],
        "responses": {
          "200": { "description": "User profile" },
          "400": { "description": "Missing user_id" },
          "404": { "description": "User not found" }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "description": "Update display_name, bio, is_private. Requires user_id query param.",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": { "type": "string", "maxLength": 50, "example": "Cool User" },
                  "bio": { "type": "string", "maxLength": 200, "example": "Living the grind" },
                  "isPrivate": { "type": "boolean", "example": false }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated profile" },
          "400": { "description": "Validation error" },
          "404": { "description": "User not found" }
        }
      }
    },
...
    "/api/users/search": {
      "get": {
        "tags": ["Friends & Social"],
        "summary": "Search users",
        "description": "Search users by username prefix. Test: /api/users/search?q=maya&user_id=YOUR_UUID",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 2 }, "description": "Username prefix (min 2 chars)" },
          { "name": "user_id", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" }, "description": "Exclude self from results" }
        ],
        "responses": {
          "200": { "description": "List of matching users", "content": { "application/json": { "schema": { "type": "object", "properties": { "users": { "type": "array", "items": { "type": "object", "properties": { "id": { "type": "string" }, "username": { "type": "string" }, "display_name": { "type": "string" }, "avatar_url": { "type": "string" }, "is_private": { "type": "boolean" } } } } } } } } },
          "400": { "description": "Query too short" }
        }
      }
    },
...
    "/api/users/{userId}/profile": {
      "get": {
        "tags": ["Friends & Social"],
        "summary": "Get user profile with privacy",
        "description": "Privacy-aware profile. Friends see full data; strangers see limited info. Test: /api/users/TARGET_UUID/profile?user_id=YOUR_UUID",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Target user ID" },
          { "name": "user_id", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" }, "description": "Current viewer ID" }
        ],
        "responses": {
          "200": { "description": "User profile with relationship-based data", "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "type": "object" }, "relationship": { "type": "string", "enum": ["self", "friend", "pending_sent", "pending_received", "squadmate", "none"] }, "stats": { "type": "object", "nullable": true }, "badges": { "type": "array", "nullable": true }, "latest_activities": { "type": "array", "nullable": true } } } } } },
          "404": { "description": "User not found" }
        }
      }
    }
  }
}